+++ CHANGELOG.md
@@ -4,11 +4,12 @@ All notable changes to the "miisync" extension will be documented in this file.
 
 Check [Keep a Changelog](http://keepachangelog.com/) for recommendations on how to structure this file.
 
-## [0.10.0] - 09/06/2023
+## [0.10.0] - 12/06/2023
 
 ### Added
 
 - `Delete Folder` explorer command.
+- New Config: `Include` which basically will override ignores if they match the same.
 - New Setting: `Request Limit`.
 - Progress window for downloading/uploading folder. Now can be cancellable by pressing it thus changing it to notification window and clicking 'Cancel'
 
@@ -22,6 +23,11 @@ Check [Keep a Changelog](http://keepachangelog.com/) for recommendations on how 
 
 - Upload folder logging "Download Folder" text.
 - Upload folder not creating empty folders.
+
+### Fixed
+
+- Where pressing on upload/download file on fileproperties view wouldn't work.
+
 
 ## [0.9.1] - 29/05/2023